make all doubles floating

+++ src/main/java/perceptronbp/Runner.java
@@ -24,7 +24,7 @@ public class Runner {
     }
 
     public static void runLetterRecognizeSuite() throws IOException {
-        double lambda = 1d;
+        float lambda = 1f;
 
         List<Layer> layers = new LayerFactory(24)
                 .addLayer(20, new TangensoidActivationFunction(lambda))
@@ -32,12 +32,12 @@ public class Runner {
                 .getLayers();
 
         Perceptron perceptron = new Perceptron.Builder(layers)
-                .withLearningCoefficient(0.1d)
+                .withLearningCoefficient(0.1f)
                 .build();
 
         // prepare trainData
-        ArrayList<double[]> trainData = new ArrayList<>();
-        ArrayList<double[]> dOutputs = new ArrayList<>();
+        ArrayList<float[]> trainData = new ArrayList<>();
+        ArrayList<float[]> dOutputs = new ArrayList<>();
 
         int filledColorCode = -14503604;
         int pixelSize = 36;
@@ -53,14 +53,14 @@ public class Runner {
         System.out.println("NOW TESTING!");
 
         // test data
-        testLetter(perceptron, inputExtractor.get("img/a6.bmp"), new double[]{1, 0, 0, 0, 0});
-        testLetter(perceptron, inputExtractor.get("img/b5.bmp"), new double[]{0, 1, 0, 0, 0});
-        testLetter(perceptron, inputExtractor.get("img/c5.bmp"), new double[]{0, 0, 1, 0, 0});
-        testLetter(perceptron, inputExtractor.get("img/d5.bmp"), new double[]{0, 0, 0, 1, 0});
-        testLetter(perceptron, inputExtractor.get("img/e5.bmp"), new double[]{0, 0, 0, 0, 1});
+        testLetter(perceptron, inputExtractor.get("img/a6.bmp"), new float[]{1, 0, 0, 0, 0});
+        testLetter(perceptron, inputExtractor.get("img/b5.bmp"), new float[]{0, 1, 0, 0, 0});
+        testLetter(perceptron, inputExtractor.get("img/c5.bmp"), new float[]{0, 0, 1, 0, 0});
+        testLetter(perceptron, inputExtractor.get("img/d5.bmp"), new float[]{0, 0, 0, 1, 0});
+        testLetter(perceptron, inputExtractor.get("img/e5.bmp"), new float[]{0, 0, 0, 0, 1});
     }
 
-    public static void testLetter(Perceptron perceptron, double[] input, double[] desiredOutput) {
+    public static void testLetter(Perceptron perceptron, float[] input, float[] desiredOutput) {
         printInputLetter(input);
         System.out.println("\r\nI think it's:");
         printOutputLetter(perceptron.calculateOutput(input));
@@ -78,40 +78,40 @@ public class Runner {
         TrainData trainData = new TrainData();
         //neurons: {A, B, C, D, E}. e.g. desired output for neuron B is {0, 1, 0, 0, 0}
         // A
-        trainData.add(inputExtractor.get("img/a1.bmp"), new double[]{1, 0, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/a2.bmp"), new double[]{1, 0, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/a3.bmp"), new double[]{1, 0, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/a4.bmp"), new double[]{1, 0, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/a5.bmp"), new double[]{1, 0, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/a1.bmp"), new float[]{1, 0, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/a2.bmp"), new float[]{1, 0, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/a3.bmp"), new float[]{1, 0, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/a4.bmp"), new float[]{1, 0, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/a5.bmp"), new float[]{1, 0, 0, 0, 0});
 
         // B
-        trainData.add(inputExtractor.get("img/b1.bmp"), new double[]{0, 1, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/b2.bmp"), new double[]{0, 1, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/b3.bmp"), new double[]{0, 1, 0, 0, 0});
-        trainData.add(inputExtractor.get("img/b4.bmp"), new double[]{0, 1, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/b1.bmp"), new float[]{0, 1, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/b2.bmp"), new float[]{0, 1, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/b3.bmp"), new float[]{0, 1, 0, 0, 0});
+        trainData.add(inputExtractor.get("img/b4.bmp"), new float[]{0, 1, 0, 0, 0});
 
         // C
-        trainData.add(inputExtractor.get("img/c1.bmp"), new double[]{0, 0, 1, 0, 0});
-        trainData.add(inputExtractor.get("img/c2.bmp"), new double[]{0, 0, 1, 0, 0});
-        trainData.add(inputExtractor.get("img/c3.bmp"), new double[]{0, 0, 1, 0, 0});
-        trainData.add(inputExtractor.get("img/c4.bmp"), new double[]{0, 0, 1, 0, 0});
+        trainData.add(inputExtractor.get("img/c1.bmp"), new float[]{0, 0, 1, 0, 0});
+        trainData.add(inputExtractor.get("img/c2.bmp"), new float[]{0, 0, 1, 0, 0});
+        trainData.add(inputExtractor.get("img/c3.bmp"), new float[]{0, 0, 1, 0, 0});
+        trainData.add(inputExtractor.get("img/c4.bmp"), new float[]{0, 0, 1, 0, 0});
 
         // D
-        trainData.add(inputExtractor.get("img/d1.bmp"), new double[]{0, 0, 0, 1, 0});
-        trainData.add(inputExtractor.get("img/d2.bmp"), new double[]{0, 0, 0, 1, 0});
-        trainData.add(inputExtractor.get("img/d3.bmp"), new double[]{0, 0, 0, 1, 0});
-        trainData.add(inputExtractor.get("img/d4.bmp"), new double[]{0, 0, 0, 1, 0});
+        trainData.add(inputExtractor.get("img/d1.bmp"), new float[]{0, 0, 0, 1, 0});
+        trainData.add(inputExtractor.get("img/d2.bmp"), new float[]{0, 0, 0, 1, 0});
+        trainData.add(inputExtractor.get("img/d3.bmp"), new float[]{0, 0, 0, 1, 0});
+        trainData.add(inputExtractor.get("img/d4.bmp"), new float[]{0, 0, 0, 1, 0});
 
         // E
-        trainData.add(inputExtractor.get("img/e1.bmp"), new double[]{0, 0, 0, 0, 1});
-        trainData.add(inputExtractor.get("img/e2.bmp"), new double[]{0, 0, 0, 0, 1});
-        trainData.add(inputExtractor.get("img/e3.bmp"), new double[]{0, 0, 0, 0, 1});
-        trainData.add(inputExtractor.get("img/e4.bmp"), new double[]{0, 0, 0, 0, 1});
+        trainData.add(inputExtractor.get("img/e1.bmp"), new float[]{0, 0, 0, 0, 1});
+        trainData.add(inputExtractor.get("img/e2.bmp"), new float[]{0, 0, 0, 0, 1});
+        trainData.add(inputExtractor.get("img/e3.bmp"), new float[]{0, 0, 0, 0, 1});
+        trainData.add(inputExtractor.get("img/e4.bmp"), new float[]{0, 0, 0, 0, 1});
 
         return trainData;
     }
 
-    private static void printInputLetter (double[] input) {
+    private static void printInputLetter (float[] input) {
         String result="";
         int count = 0;
         for (int i = 0; i < 6; ++i){
@@ -127,8 +127,8 @@ public class Runner {
         System.out.println(result);
     }
 
-    private static void printOutputLetter (double[] input) {
-        double maxValue = input[0];
+    private static void printOutputLetter (float[] input) {
+        float maxValue = input[0];
         int maxValueIndex = 0;
 
         for (int i = 1; i < input.length; ++i) {
@@ -146,7 +146,7 @@ public class Runner {
     public static void runXorSuite() {
         // implementing XOR function
         //2 inputs, 2 hidden layers, 1 output layer. Can store any number of hidden layers
-        double lambda = 1d;
+        float lambda = 1f;
 
         List<Layer> layers2 = new LayerFactory(2)
                 .addLayer(20, new TangensoidActivationFunction(lambda))
@@ -155,7 +155,7 @@ public class Runner {
                 .getLayers();
 
         Perceptron perceptron2 =  new Perceptron.Builder(layers2)
-                .withLearningCoefficient(0.01d)
+                .withLearningCoefficient(0.01f)
                 .build();
 
         // learning
@@ -163,22 +163,22 @@ public class Runner {
         perceptron2.learn(createTrainDataXOR(), maxEpochs);
 
         // testing
-        System.out.println("Should be false: " + testXor(perceptron2.calculateOutput(new double[]{0,0}), 0.1));
-        System.out.println("Should be true: " + testXor(perceptron2.calculateOutput(new double[]{0,1}), 0.1));
-        System.out.println("Should be true: " + testXor(perceptron2.calculateOutput(new double[]{1,0}), 0.1));
-        System.out.println("Should be false: " + testXor(perceptron2.calculateOutput(new double[]{0,0}), 0.1));
+        System.out.println("Should be false: " + testXor(perceptron2.calculateOutput(new float[]{0,0}), 0.1f));
+        System.out.println("Should be true: " + testXor(perceptron2.calculateOutput(new float[]{0,1}), 0.1f));
+        System.out.println("Should be true: " + testXor(perceptron2.calculateOutput(new float[]{1,0}), 0.1f));
+        System.out.println("Should be false: " + testXor(perceptron2.calculateOutput(new float[]{0,0}), 0.1f));
 
     }
-    public static boolean testXor(double[] value, double threshold) {
+    public static boolean testXor(float[] value, float threshold) {
         return value[0] > threshold;
     }
 
     public static TrainData createTrainDataXOR() {
         List<InputAndDesiredOutput> inputAndDesiredOutputList = new ArrayList<>(4);
-        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new double[]{0,0}, new double[]{0}));
-        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new double[]{0,1}, new double[]{1}));
-        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new double[]{1,0}, new double[]{1}));
-        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new double[]{1,1}, new double[]{0}));
+        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new float[]{0,0}, new float[]{0}));
+        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new float[]{0,1}, new float[]{1}));
+        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new float[]{1,0}, new float[]{1}));
+        inputAndDesiredOutputList.add(new InputAndDesiredOutput(new float[]{1,1}, new float[]{0}));
 
         TrainData trainData = new TrainData(inputAndDesiredOutputList);
